QCamera: Additional interface cleanup
- Synced changes in 'QCamera2/stack'
  with master branch.
- Also fixed type conversion issues.

CRs-Fixed: 641033

+++ QCamera2/HAL/QCameraPostProc.cpp
@@ -139,7 +139,7 @@ int32_t QCameraPostProcessor::init(jpeg_encode_callback_t jpeg_cb, void *user_da
 
     if ((0 > m_parent->m_max_pic_width) || (0 > m_parent->m_max_pic_height)) {
         ALOGE("%s : Negative dimension %dx%d", __func__,
-        m_parent->m_max_pic_width, m_parent->m_max_pic_height);
+                m_parent->m_max_pic_width, m_parent->m_max_pic_height);
         return BAD_VALUE;
     }
 
@@ -462,7 +462,7 @@ int32_t QCameraPostProcessor::getJpegEncodingConfig(mm_jpeg_encode_params_t& enc
         goto on_error;
     }
     encode_parm.num_src_bufs = pStreamMem->getCnt();
-    for (uint32_t i = 0; i < (uint32_t)encode_parm.num_src_bufs; i++) {
+    for (uint32_t i = 0; i < encode_parm.num_src_bufs; i++) {
         camera_memory_t *stream_mem = pStreamMem->getMemory(i, false);
         if (stream_mem != NULL) {
             encode_parm.src_main_buf[i].index = i;
@@ -713,7 +713,7 @@ int32_t QCameraPostProcessor::processData(mm_camera_super_buf_t *frame)
 
         // find meta data frame
         mm_camera_buf_def_t *meta_frame = NULL;
-        for (int i = 0; i < frame->num_bufs; i++) {
+        for (uint32_t i = 0; i < frame->num_bufs; i++) {
             // look through input superbuf
             if (frame->bufs[i]->stream_type == CAM_STREAM_TYPE_METADATA) {
                 meta_frame = frame->bufs[i];
@@ -971,7 +971,7 @@ int32_t QCameraPostProcessor::processPPData(mm_camera_super_buf_t *frame)
 
     // find meta data frame
     mm_camera_buf_def_t *meta_frame = NULL;
-    for (int i = 0; job && (i < job->src_frame->num_bufs); i++) {
+    for (uint32_t i = 0; job && (i < job->src_frame->num_bufs); i++) {
         // look through input superbuf
         if (job->src_frame->bufs[i]->stream_type == CAM_STREAM_TYPE_METADATA) {
             meta_frame = job->src_frame->bufs[i];
@@ -981,7 +981,7 @@ int32_t QCameraPostProcessor::processPPData(mm_camera_super_buf_t *frame)
 
     if (meta_frame == NULL) {
         // look through reprocess superbuf
-        for (int i = 0; i < frame->num_bufs; i++) {
+        for (uint32_t i = 0; i < frame->num_bufs; i++) {
             if (frame->bufs[i]->stream_type == CAM_STREAM_TYPE_METADATA) {
                 meta_frame = frame->bufs[i];
                 break;
@@ -1407,7 +1407,7 @@ int32_t QCameraPostProcessor::queryStreams(QCameraStream **main,
     *main = *thumb = NULL;
     *main_image = *thumb_image = NULL;
     // find snapshot frame and thumnail frame
-    for (int i = 0; i < frame->num_bufs; i++) {
+    for (uint32_t i = 0; i < frame->num_bufs; i++) {
         QCameraStream *pStream =
                 pChannel->getStreamByHandle(frame->bufs[i]->stream_id);
         if (pStream != NULL) {
@@ -1430,7 +1430,7 @@ int32_t QCameraPostProcessor::queryStreams(QCameraStream **main,
         pSrcReprocChannel = m_parent->getChannelByHandle(reproc_frame->ch_id);
         if (pSrcReprocChannel != NULL) {
             // find thumbnail frame
-            for (int i = 0; i < reproc_frame->num_bufs; i++) {
+            for (uint32_t i = 0; i < reproc_frame->num_bufs; i++) {
                 QCameraStream *pStream =
                         pSrcReprocChannel->getStreamByHandle(
                                 reproc_frame->bufs[i]->stream_id);
@@ -1914,7 +1914,7 @@ int32_t QCameraPostProcessor::processRawImageImpl(mm_camera_super_buf_t *recvd_f
     }
 
     // find snapshot frame
-    for (int i = 0; i < recvd_frame->num_bufs; i++) {
+    for (uint32_t i = 0; i < recvd_frame->num_bufs; i++) {
         QCameraStream *pCurStream =
             pChannel->getStreamByHandle(recvd_frame->bufs[i]->stream_id);
         if (pCurStream != NULL) {
@@ -2407,29 +2407,28 @@ int32_t QCameraPostProcessor::reprocess(qcamera_pp_data_t *pp_job)
     }
 
     if (m_parent->isRegularCapture()) {
-       if ((NULL != pp_job->src_frame) &&
-           (0 < pp_job->src_frame->num_bufs)) {
-           mm_camera_buf_def_t *bufs = NULL;
-           uint8_t num_bufs = pp_job->src_frame->num_bufs;
-           bufs = new mm_camera_buf_def_t[num_bufs];
-           if (NULL == bufs) {
-               ALOGE("%s:Unable to allocate cached buffers",
-                     __func__);
-               return NO_MEMORY;
-           }
+        if ((NULL != pp_job->src_frame) &&
+                (0 < pp_job->src_frame->num_bufs)) {
+            mm_camera_buf_def_t *bufs = NULL;
+            uint32_t num_bufs = pp_job->src_frame->num_bufs;
+            bufs = new mm_camera_buf_def_t[num_bufs];
+            if (NULL == bufs) {
+                ALOGE("%s:Unable to allocate cached buffers", __func__);
+                return NO_MEMORY;
+            }
 
-           for (int i = 0; i < num_bufs; i++) {
-               bufs[i] = *pp_job->src_frame->bufs[i];
-               pp_job->src_frame->bufs[i] = &bufs[i];
-           }
-           pp_job->src_reproc_bufs = bufs;
-       }
+            for (uint32_t i = 0; i < num_bufs; i++) {
+                bufs[i] = *pp_job->src_frame->bufs[i];
+                pp_job->src_frame->bufs[i] = &bufs[i];
+            }
+            pp_job->src_reproc_bufs = bufs;
+        }
 
-       // Don't release source frame after encoding
-       // at this point the source channel will not exist.
-       pp_job->reproc_frame_release = true;
-       m_ongoingPPQ.enqueue((void *)pp_job);
-       rc = pChannel->doReprocessOffline(pp_job->src_frame);
+        // Don't release source frame after encoding
+        // at this point the source channel will not exist.
+        pp_job->reproc_frame_release = true;
+        m_ongoingPPQ.enqueue((void *)pp_job);
+        rc = pChannel->doReprocessOffline(pp_job->src_frame);
     } else {
         m_ongoingPPQ.enqueue((void *)pp_job);
         rc = pChannel->doReprocess(pp_job->src_frame);
@@ -2521,7 +2520,7 @@ int32_t QCameraPostProcessor::setYUVFrameInfo(mm_camera_super_buf_t *recvd_frame
     }
 
     // find snapshot frame
-    for (int i = 0; i < recvd_frame->num_bufs; i++) {
+    for (uint32_t i = 0; i < recvd_frame->num_bufs; i++) {
         QCameraStream *pStream =
             pChannel->getStreamByHandle(recvd_frame->bufs[i]->stream_id);
         if (pStream != NULL) {
